+++ daglib/__init__.py
@@ -1,4 +1,5 @@
-from daglib.dag import Dag
-from daglib.metadata import MetaDB
+from daglib.core.dag import Dag
+from daglib.core.metadata import MetaDB
+from daglib.core.assets import asset
 
 __version__ = "0.3.0"
